Return 404 On Change Of Not Assigned Versions And More:
This is a little chaotic commit. I noticed, that I used the wrong parent classes in tests,
and the corrections brought some errors to light: POST was checked in the wrong method.
Returning 403 on not assigned LemmaVersion.LemmaArticles, became to extensive
and 404 maybe a better response anyway.

Changes:
- modified:   permissions.py
- modified:   tests/apitests/test_lemma_version.py

Tags: [skip ci]